Move aragonjs-wrapper to helpers folder
