+++ packages/aragon-cli/src/commands/dao_cmds/acl_cmds/view.js
@@ -1,4 +1,4 @@
-import { initAragonJS, getApps } from '../utils/aragonjs-wrapper'
+import { initAragonJS, getApps } from '../../../helpers/aragonjs-wrapper'
 const chalk = require('chalk')
 const TaskList = require('listr')
 const daoArg = require('../utils/daoArg')
